NNTraining_ImageProcessing
Using images from the web, create a dataset with at least 2500 samples,
for training a neural network to recognize if an input image is a face /
non-face. Split the dataset into training, validation, and test sets.
Train a neural network using (a) hinge loss function (b) cross entropy loss
function. Use at least 2 hidden layers. Plot the training, validation, and
test loss values across epochs.
Determine the receptive field of any chosen neuron in each layer. Explain
your approach to determining this. The receptive field of a neuron is defined as the primary (first) layer input that excites this neuron the most.
Add a regularization term and determine how the training changes. Determine how to tune the hyperparameter that controls the emphasis on
the mis-classification term.