Out of Sample Learning-On Regression Approaches
Build  3 regression approaches, of which kernel SVR, and neural network based regression, are two.

Generate datasets using known functions of 1, 2, and over 6 variables.
In the training data, leave out specific regions from the input. For example, if f(x)=x2
, you can leave out training data in the region −0.5≤x≤0.5
 or any other subrange. You can choose similar regions for higher dimensional functions.
Train the regressors and test their performance in the regions missed out of training.
Suggest, and demonstrate how the performance can be improved.